It appears that only question checkboxes have popup text, and they only show it when checked.
