+++ src/components/shared/questions/index.js
@@ -1,6 +1,7 @@
 export * from './question_check_box_field'
 export * from './question_field'
 export * from './question_label'
+export * from './question_popup'
 export * from './question_radio_button_field'
 export * from './question_select_field'
 export * from "./question_text_area_field"
